Serialise inexactness parameters of LinMoreAlgorithm
Although these are less likely to be used, we need to at least
serialise the useInexact_ vector correctly, otherwise we'll read
uninitialised data during one of the checks when the algorithm is
run. Because of this, the call site for LinMoreAlgorithm serialisation
also needs to include cereal/types/vector.hpp to work with std::vector.